chore: Move nearby_distance_meter to heuristic
