+++ setup.py
@@ -98,11 +98,16 @@ setup(
         'License :: OSI Approved :: Apache Software License',
         'Operating System :: OS Independent'
     ],
-    packages=['timefold.solver', 'timefold.solver.domain',
-              'timefold.solver.config', 'timefold.solver.score',
+    packages=['timefold.solver',
+              'timefold.solver.config',
+              'timefold.solver.domain',
+              'timefold.solver.heuristic',
+              'timefold.solver.score',
               'timefold.solver.test',
               'jpyinterpreter',
-              'java-stubs', 'jpype-stubs', 'ai-stubs'],
+              'java-stubs',
+              'jpype-stubs',
+              'ai-stubs'],
     package_dir={
         'timefold.solver': 'timefold-solver-python-core/src/main/python',
         'jpyinterpreter': 'jpyinterpreter/src/main/python',
