+++ timefold-solver-python-core/src/main/python/heuristic/__init__.py
@@ -1,0 +1,1 @@
+from ._nearby_selection import *
